requirements.txt
python-telegram-bot
requests
google-generativeai
python-dotenv
schedule
